implement dashboard feature

+++ src/main/java/lk/ashan/routenetlkserverapllication/report/model/projection/Report2Projection.java
@@ -1,7 +1,9 @@
 package lk.ashan.routenetlkserverapllication.report.model.projection;
 
+import java.math.BigDecimal;
+
 public interface Report2Projection {
     String getDepotName();
-    Double getCashAmount();
-    Double getDigitalAmount();
+    BigDecimal getCashAmount();
+    BigDecimal getDigitalAmount();
 }
